Refactor AWS module
