+++ libcloudforensics/providers/aws/forensics.py
@@ -114,9 +114,9 @@ def CreateVolumeCopy(zone: str,
 
   try:
     if volume_id:
-      volume_to_copy = source_account.GetVolumeById(volume_id)
+      volume_to_copy = source_account.ebs.GetVolumeById(volume_id)
     elif instance_id:
-      instance = source_account.GetInstanceById(instance_id)
+      instance = source_account.ec2.GetInstanceById(instance_id)
       volume_to_copy = instance.GetBootVolume()
 
     if not volume_type:
@@ -127,8 +127,8 @@ def CreateVolumeCopy(zone: str,
     snapshot = volume_to_copy.Snapshot()
     logger.info('Created snapshot: {0:s}'.format(snapshot.snapshot_id))
 
-    source_account_id = source_account.GetAccountInformation('Account')
-    destination_account_id = destination_account.GetAccountInformation(
+    source_account_id = source_account.ebs.GetAccountInformation('Account')
+    destination_account_id = destination_account.ebs.GetAccountInformation(
         'Account')
 
     if source_account_id != destination_account_id:
@@ -140,8 +140,8 @@ def CreateVolumeCopy(zone: str,
             'Encrypted volume detected, generating one-time use CMK key')
         # Generate one-time use KMS key that will be shared with the
         # destination account.
-        kms_key_id = source_account.CreateKMSKey()
-        source_account.ShareKMSKeyWithAWSAccount(
+        kms_key_id = source_account.kms.CreateKMSKey()
+        source_account.kms.ShareKMSKeyWithAWSAccount(
             kms_key_id, destination_account_id)
         # Create a copy of the initial snapshot and encrypts it with the
         # shared key
@@ -158,13 +158,13 @@ def CreateVolumeCopy(zone: str,
       snapshot = snapshot.Copy(delete=True, deletion_account=source_account)
 
     if tags and tags.get('Name'):
-      new_volume = destination_account.CreateVolumeFromSnapshot(
+      new_volume = destination_account.ebs.CreateVolumeFromSnapshot(
           snapshot,
           volume_type=volume_type,
           volume_name=tags['Name'],
           tags=tags)
     else:
-      new_volume = destination_account.CreateVolumeFromSnapshot(
+      new_volume = destination_account.ebs.CreateVolumeFromSnapshot(
           snapshot,
           volume_type=volume_type,
           volume_name_prefix='evidence',
@@ -176,7 +176,7 @@ def CreateVolumeCopy(zone: str,
 
     snapshot.Delete()
     # Delete the one-time use KMS key, if one was generated
-    source_account.DeleteKMSKey(kms_key_id)
+    source_account.kms.DeleteKMSKey(kms_key_id)
     logger.info('Done')
   except RuntimeError as exception:
     error_msg = 'Copying volume {0:s}: {1!s}'.format(
@@ -249,7 +249,7 @@ def StartAnalysisVm(
   if not ami:
     logger.info('No AMI provided, fetching one for Ubuntu 18.04')
     qfilter = [{'Name': 'name', 'Values': [UBUNTU_1804_FILTER]}]
-    ami_list = aws_account.ListImages(qfilter)
+    ami_list = aws_account.ec2.ListImages(qfilter)
     # We should only get 1 AMI image back, if we get multiple we
     # have no way of knowing which one to use.
     if len(ami_list) > 1:
@@ -260,7 +260,7 @@ def StartAnalysisVm(
   assert ami  # Mypy: assert that ami is not None
 
   logger.info('Starting analysis VM {0:s}'.format(vm_name))
-  analysis_vm, created = aws_account.GetOrCreateAnalysisVm(
+  analysis_vm, created = aws_account.ec2.GetOrCreateAnalysisVm(
       vm_name,
       boot_volume_size,
       ami,
@@ -272,6 +272,7 @@ def StartAnalysisVm(
   for volume_id, device_name in (attach_volumes or []):
     logger.info('Attaching volume {0:s} to device {1:s}'.format(
         volume_id, device_name))
-    analysis_vm.AttachVolume(aws_account.GetVolumeById(volume_id), device_name)
+    analysis_vm.AttachVolume(
+        aws_account.ebs.GetVolumeById(volume_id), device_name)
   logger.info('VM ready.')
   return analysis_vm, created
