+++ libcloudforensics/providers/aws/internal/ebs.py
@@ -13,8 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Disk functionality."""
-
-from typing import TYPE_CHECKING, Dict, Optional, Union
+import binascii
+from typing import TYPE_CHECKING, Dict, Optional, Union, List, Any
 
 import botocore
 
@@ -305,3 +305,308 @@ class AWSSnapshot(AWSElasticBlockStore):
         },
         OperationType='add'
     )
+
+
+class AWSEBS:
+  """Class that represents AWS EBS storage services."""
+
+  def __init__(self,
+               aws_account: 'account.AWSAccount') -> None:
+    """Initialize the AWS ebs client object.
+
+    Args:
+      aws_account (AWSAccount): An AWS account object.
+    """
+    self.aws_account = aws_account
+
+  def ListVolumes(
+      self,
+      region: Optional[str] = None,
+      filters: Optional[List[Dict[str, Any]]] = None) -> Dict[str, AWSVolume]:
+    """List volumes of an AWS account.
+
+    Example usage:
+      # List volumes attached to the instance 'some-instance-id'
+      ListVolumes(filters=[
+          {'Name':'attachment.instance-id', 'Values':['some-instance-id']}])
+
+    Args:
+      region (str): Optional. The region from which to list the volumes.
+          If none provided, the default_region associated to the AWSAccount
+          object will be used.
+      filters (List[Dict]): Optional. Filters for the query. Filters are
+          given as a list of dictionaries, e.g.: {'Name': 'someFilter',
+          'Values': ['value1', 'value2']}.
+
+    Returns:
+      Dict[str, AWSVolume]: Dictionary mapping volume IDs (str) to their
+          respective AWSVolume object.
+
+    Raises:
+      RuntimeError: If volumes can't be listed.
+    """
+
+    if not filters:
+      filters = []
+
+    volumes = {}
+    client = self.aws_account.ClientApi(common.EC2_SERVICE, region=region)
+    responses = common.ExecuteRequest(
+        client, 'describe_volumes', {'Filters': filters})
+    for response in responses:
+      for volume in response['Volumes']:
+        volume_id = volume['VolumeId']
+        aws_volume = AWSVolume(volume_id,
+                               self.aws_account,
+                               self.aws_account.default_region,
+                               volume['AvailabilityZone'],
+                               volume['Encrypted'])
+
+        for tag in volume.get('Tags', []):
+          if tag.get('Key') == 'Name':
+            aws_volume.name = tag.get('Value')
+            break
+
+        for attachment in volume.get('Attachments', []):
+          if attachment.get('State') == 'attached':
+            aws_volume.device_name = attachment.get('Device')
+            break
+
+        volumes[volume_id] = aws_volume
+    return volumes
+
+  def GetVolumesByNameOrId(self,
+                           volume_name: Optional[str] = None,
+                           volume_id: Optional[str] = None,
+                           region: Optional[str] = None) -> List[AWSVolume]:
+    """Get a volume from an AWS account by its name tag or its ID.
+
+    Exactly one of [volume_name, volume_id] must be specified. If looking up
+    a volume by its ID, the method returns a list with exactly one
+    element. If looking up volumes by their name tag (which are not unique
+    across volumes), then the method will return a list of all volumes
+    with that name tag, or an empty list if no volumes with matching name tag
+    could be found.
+
+    Args:
+      volume_name (str): Optional. The volume name tag of the volume to get.
+      volume_id (str): Optional. The volume id of the volume to get.
+      region (str): Optional. The region to look the volume in.
+          If none provided, the default_region associated to the AWSAccount
+          object will be used.
+
+    Returns:
+      List[AWSVolume]: A list of Amazon EC2 Volume objects.
+
+    Raises:
+      ValueError: If both volume_name and volume_id are None or if both
+          are set.
+    """
+
+    if (not volume_name and not volume_id) or (volume_name and volume_id):
+      raise ValueError('You must specify exactly one of [volume_name, '
+                       'volume_id]. Got volume_name: {0:s}, volume_id: '
+                       '{1:s}'.format(str(volume_name), str(volume_id)))
+    if volume_name:
+      return self.GetVolumesByName(volume_name, region=region)
+    assert volume_id  # Mypy: assert that volume_id is not None
+    return [self.GetVolumeById(volume_id, region=region)]
+
+  def GetVolumesByName(self,
+                       volume_name: str,
+                       region: Optional[str] = None) -> List[AWSVolume]:
+    """Get all volumes from an AWS account with matching name tag.
+
+    Args:
+      volume_name (str): The volume name tag.
+      region (str): Optional. The region to look the volume in.
+          If none provided, the default_region associated to the AWSAccount
+          object will be used.
+
+    Returns:
+      List[AWSVolume]: A list of EC2 Volume objects. If no volume with
+          matching name tag is found, the method returns an empty list.
+    """
+
+    matching_volumes = []
+    volumes = self.ListVolumes(region=region)
+    for volume_id in volumes:
+      volume = volumes[volume_id]
+      if volume.name == volume_name:
+        matching_volumes.append(volume)
+    return matching_volumes
+
+  def GetVolumeById(self,
+                    volume_id: str,
+                    region: Optional[str] = None) -> AWSVolume:
+    """Get a volume from an AWS account by its ID.
+
+    Args:
+      volume_id (str): The volume id.
+      region (str): Optional. The region to look the volume in.
+          If none provided, the default_region associated to the AWSAccount
+          object will be used.
+
+    Returns:
+      AWSVolume: An Amazon EC2 Volume object.
+
+    Raises:
+      RuntimeError: If volume does not exist.
+    """
+
+    volumes = self.ListVolumes(region=region)
+    volume = volumes.get(volume_id)
+    if not volume:
+      error_msg = 'Volume {0:s} was not found in AWS account'.format(
+          volume_id)
+      raise RuntimeError(error_msg)
+    return volume
+
+  def CreateVolumeFromSnapshot(
+      self,
+      snapshot: AWSSnapshot,
+      volume_name: Optional[str] = None,
+      volume_name_prefix: Optional[str] = None,
+      volume_type: str = 'gp2',
+      kms_key_id: Optional[str] = None,
+      tags: Optional[Dict[str, str]] = None) -> AWSVolume:
+    """Create a new volume based on a snapshot.
+
+    Args:
+      snapshot (AWSSnapshot): Snapshot to use.
+      volume_name (str): Optional. String to use as new volume name.
+      volume_name_prefix (str): Optional. String to prefix the volume name with.
+      volume_type (str): Optional. The volume type for the volume to create.
+          Can be one of 'standard'|'io1'|'gp2'|'sc1'|'st1'. The default is
+          'gp2'.
+      kms_key_id (str): Optional. A KMS key id to encrypt the volume with.
+      tags (Dict[str, str]): Optional. A dictionary of tags to add to the
+          volume, for example {'TicketID': 'xxx'}. An entry for the volume
+          name is added by default.
+
+    Returns:
+      AWSVolume: An AWS EBS Volume.
+
+    Raises:
+      ValueError: If the volume name does not comply with the RegEx,
+          or if the volume type is invalid.
+      RuntimeError: If the volume could not be created.
+    """
+
+    if volume_type not in ['standard', 'io1', 'gp2', 'sc1', 'st1']:
+      raise ValueError('Volume type must be one of [standard, io1, gp2, sc1, '
+                       'st1]. Got: {0:s}'.format(volume_type))
+
+    if not volume_name:
+      volume_name = self._GenerateVolumeName(
+          snapshot, volume_name_prefix=volume_name_prefix)
+
+    if len(volume_name) > 255:
+      raise ValueError(
+          'Volume name {0:s} is too long (>255 chars)'.format(volume_name))
+
+    if not tags:
+      tags = {}
+    tags['Name'] = volume_name
+
+    client = self.aws_account.ClientApi(common.EC2_SERVICE)
+    create_volume_args = {
+        'AvailabilityZone': snapshot.availability_zone,
+        'SnapshotId': snapshot.snapshot_id,
+        'TagSpecifications': [common.CreateTags(common.VOLUME, tags)],
+        'VolumeType': volume_type
+    }
+    if kms_key_id:
+      create_volume_args['Encrypted'] = True
+      create_volume_args['KmsKeyId'] = kms_key_id
+    if volume_type == 'io1':
+      # If using the io1 volume type, we must specify Iops, see
+      # https://boto3.amazonaws.com/v1/documentation/api/latest/reference/
+      # services/ec2.html#EC2.Client.create_volume. io1 volumes allow for a
+      # ratio of 50 IOPS per 1 GiB.
+      create_volume_args['Iops'] = self.aws_account.ResourceApi(
+          common.EC2_SERVICE).Snapshot(snapshot.snapshot_id).volume_size * 50
+    try:
+      volume = client.create_volume(**create_volume_args)
+      volume_id = volume['VolumeId']
+      # Wait for volume creation completion
+      client.get_waiter('volume_available').wait(VolumeIds=[volume_id])
+    except (client.exceptions.ClientError,
+            botocore.exceptions.WaiterError) as exception:
+      raise RuntimeError('Could not create volume {0:s} from snapshot '
+                         '{1:s}: {2:s}'.format(volume_name, snapshot.name,
+                                               str(exception)))
+
+    zone = volume['AvailabilityZone']
+    encrypted = volume['Encrypted']
+
+    return AWSVolume(volume_id,
+                     self.aws_account,
+                     self.aws_account.default_region,
+                     zone,
+                     encrypted,
+                     name=volume_name)
+
+  def GetAccountInformation(self, info: str) -> str:
+    """Get information about the AWS account in use.
+
+    If the call succeeds, then the response from the STS API is expected to
+    have the following entries:
+      - UserId
+      - Account
+      - Arn
+    See https://boto3.amazonaws.com/v1/documentation/api/1.9.42/reference/services/sts.html#STS.Client.get_caller_identity for more details. # pylint: disable=line-too-long
+
+    Args:
+      info (str): The account information to retrieve. Must be one of [UserID,
+          Account, Arn]
+    Returns:
+      str: The information requested.
+
+    Raises:
+      KeyError: If the requested information doesn't exist.
+    """
+    account_information = self.aws_account.ClientApi(
+        common.ACCOUNT_SERVICE).get_caller_identity()  # type: Dict[str, str]
+    if not account_information.get(info):
+      raise KeyError('Key must be one of ["UserId", "Account", "Arn"]')
+    return account_information[info]
+
+  def _GenerateVolumeName(self,
+                          snapshot: AWSSnapshot,
+                          volume_name_prefix: Optional[str] = None) -> str:
+    """Generate a new volume name given a volume's snapshot.
+
+    Args:
+      snapshot (AWSSnapshot): A volume's Snapshot.
+      volume_name_prefix (str): Optional. Prefix for the volume name.
+
+    Returns:
+      str: A name for the volume.
+
+    Raises:
+      ValueError: If the volume name does not comply with the RegEx.
+    """
+
+    # Max length of tag values in AWS is 255 characters
+    user_id = self.GetAccountInformation('UserId')
+    volume_id = user_id + snapshot.volume.volume_id
+    volume_id_crc32 = '{0:08x}'.format(
+        binascii.crc32(volume_id.encode()) & 0xffffffff)
+    truncate_at = 255 - len(volume_id_crc32) - len('-copy') - 1
+    assert snapshot.name  # Mypy: assert that snapshot.name is not None
+    if volume_name_prefix:
+      volume_name_prefix += '-'
+      if len(volume_name_prefix) > truncate_at:
+        # The volume name prefix is too long
+        volume_name_prefix = volume_name_prefix[:truncate_at]
+      truncate_at -= len(volume_name_prefix)
+      volume_name = '{0:s}{1:s}-{2:s}-copy'.format(
+          volume_name_prefix,
+          snapshot.name[:truncate_at],
+          volume_id_crc32)
+    else:
+      volume_name = '{0:s}-{1:s}-copy'.format(
+          snapshot.name[:truncate_at], volume_id_crc32)
+
+    return volume_name
