+++ tests/providers/aws/aws_e2e.py
@@ -153,7 +153,7 @@ class EndToEndTest(unittest.TestCase):
 
     aws_account = account.AWSAccount(self.zone)
     qfilter = [{'Name': 'name', 'Values': ['Ubuntu 18.04*']}]
-    images = aws_account.ListImages(qfilter)
+    images = aws_account.ec2.ListImages(qfilter)
 
     self.assertGreater(len(images), 0)
     self.assertIn('Name', images[0])
